Implement high score saving/showing
The user has now the ability to store their highscore at the end of a
game. If the score is high enough to be in the top 10 (configurable
number), they're asked at whether they want to save it.

Option -s is used to show the list of highscores and quit (no game).